fix: duplicate TracingService instances without configuration
By directly importing the TracingCoreModule in other modules in the package
the TracingService gets instantiated multiple times.

Only with TracingCoreModule.forRoot() the configuration is properly set,
so these additional instantiations will use default values for everything.

Because we only have one singleton instance of XRayClient and write
configuration to it, the user-desired configuration gets overriden by this
broken behaviour.